refractor: import config data into a constants.js file
